src/requirements.txt
amocrm-api==0.8.3
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
gunicorn==19.9.0
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-vision==0.37.0
google-cloud-storage==1.16.1
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pdf2image==1.5.4
Pillow==6.0.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
responses==0.10.6
rsa==4.0
six==1.12.0
transliterate==1.10.2
urllib3==1.25.3
Werkzeug==0.15.4
Wand==0.5.4
